+++ src/views/categoryCenter/official/index.vue
@@ -1,82 +1,38 @@
 <template>
   <basic-container>
-    <el-form inline :model="tempSearch" class="demo-form-inline">
-      <el-form-item label="栏目名称：">
-        <el-input v-model="tempSearch.officialColumnName" placeholder="请输入栏目名称"></el-input>
-      </el-form-item>
-      <el-form-item>
-        <el-button type="primary" @click="toSearch()">查询</el-button>
-      </el-form-item>
-      <div>
-        <el-button
-          class="filter-item"
-          type="primary"
-          size="mini"
-          icon="el-icon-plus"
-          @click="handleCreate">添加
-        </el-button>
-      </div>
-    </el-form>
-
-    <hc-table-data-box :empty="!columnList || columnList.length == 0" :loading="boxLoading">
-      <div class="column-box">
-        <div v-for="column in columnList" :key="column.officialColumnId" class="column-item">
-          <div class="column-item-info">
-            <div class="column-item-name">{{column.officialColumnName}}</div>
-            <!-- <div class="column-item-sort" v-if="column.isOpening && column.sort">No.{{column.sort}}</div> -->
-          </div>
-          <div class="column-item-option">
-            <div class="column-item-option-left">
-              <el-button v-if="confAuthority(column)" type="text" size="mini" @click="cityView(column.officialColumnId)">查看配置城市</el-button>
-              <el-button v-else-if="column.closeAllowed == '0'" type="text" size="mini" @click="handleStart(column)">{{column.havEnable ? '启用' : '停用'}}</el-button>
+    <hc-table-form title="官方发布栏目">
+      <hc-crud ref="hcCrud" :option="tableOption" :fetchListFun="fetchListFun" :addFun="addFun" :updateFun="updateFun">
+        <template v-slot:table="scope">
+           <hc-table-data-box :empty="!scope.tableData || scope.tableData.length == 0" :loading="boxLoading">
+            <div class="column-box">
+              <div v-for="column in scope.tableData" :key="column.officialColumnId" class="column-item">
+                <div class="column-item-info">
+                  <div class="column-item-name">{{column.officialColumnName}}</div>
+                  <!-- <div class="column-item-sort" v-if="column.isOpening && column.sort">No.{{column.sort}}</div> -->
+                </div>
+                <div class="column-item-option">
+                  <div class="column-item-option-left">
+                    <el-button v-if="confAuthority(column)" type="text" size="mini" @click="cityView(column.officialColumnId)">查看配置城市</el-button>
+                    <el-button v-else-if="column.closeAllowed == '0'" type="text" size="mini" @click="handleStart(column)">{{column.havEnable ? '启用' : '停用'}}</el-button>
+                  </div>
+                  <div class="column-item-option-right">
+                    <!-- <el-button v-if="!isAdmin" type="text" size="mini" @click="handleSort(column)">排序</el-button> -->
+                    <template v-if="operAuthority(column)">
+                      <el-button type="text" size="mini" @click="handleUpdate(column)">编辑</el-button>
+                      <el-button type="text" size="mini" @click="handleDel(column.officialColumnId)">删除</el-button>
+                    </template>
+                  </div>
+                </div>
+              </div>
             </div>
-            <div class="column-item-option-right">
-              <!-- <el-button v-if="!isAdmin" type="text" size="mini" @click="handleSort(column)">排序</el-button> -->
-              <template v-if="operAuthority(column)">
-                <el-button type="text" size="mini" @click="handleUpdate(column)">编辑</el-button>
-                <el-button type="text" size="mini" @click="handleDel(column.officialColumnId)">删除</el-button>
-              </template>
-            </div>
-          </div>
-        </div>
-      </div>
-    </hc-table-data-box>
+          </hc-table-data-box>
+        </template>
+      </hc-crud>
 
-    <div class="pagination-box">
-      <el-pagination
-        style="display: inline-block"
-        @size-change="sizeChange"
-        @current-change="currentChange"
-        :current-page="page.currentPage"
-        :page-sizes="[10, 20, 30,, 40, 50, 100]"
-        background
-        :page-size="page.pageSize"
-        layout="total, sizes, prev, pager, next, jumper"
-        :total="page.total">
-      </el-pagination>
-    </div>
+    </hc-table-form>
 
     <hc-city-box ref="hcCityBox"></hc-city-box>
     
-    <el-dialog
-      :title="formTitle"
-      :visible.sync="formDialogVisible"
-      width="70%">
-      <el-form :model="formData" labelWidth="150px">
-        <el-form-item label="栏目名称：">
-          <el-input v-model="formData.officialColumnName"></el-input>
-        </el-form-item>
-        <el-form-item v-if="userType == 1 || userType == 2" label="是否允许城市停用：">
-          <el-switch v-model="formData.closeAllowed" active-value="0" active-text="允许" inactive-text="不允许" inactive-value="1"></el-switch>
-        </el-form-item>
-        
-      </el-form>
-      <div slot="footer">
-        <el-button v-show="formType == 'add'" type="primary" @click="create">保 存</el-button>
-        <el-button v-show="formType == 'edit'" type="primary" @click="update">修 改</el-button>
-        <el-button @click="formDialogVisible = false">取 消</el-button>
-      </div>
-    </el-dialog>
   </basic-container>
 </template>
 
@@ -88,10 +44,6 @@ export default {
   components: { HcCityBox },
   data () {
     return {
-      tempSearch: {
-        officialColumnName: ''
-      },
-      searchForm: {},
       formData: {},
       formType: 'add',
       formDialogVisible: false,
@@ -103,7 +55,31 @@ export default {
       },
       allCityList: [],
       initCityList: [],
-      boxLoading: false
+      boxLoading: false,
+      tableOption: {
+        menu: ['add'],
+        labelWidth: '150px',
+        columns: [
+          {
+            label: '栏目名称',
+            prop: 'officialColumnName',
+            maxlength: 50,
+            search: true,
+            rules: [{required: true,message: '请输入栏目名称', trigger: 'blur'}]
+          },
+          {
+            label: '是否允许城市停用',
+            prop: 'closeAllowed',
+            type: 'switch',
+            inactiveText: '不允许',
+            inactiveValue: '1',
+            activeText: '允许',
+            activeValue: '0',
+            value: '1',
+            rules: [{required: true,message: '请选择是否允许城市停用', trigger: 'blur'}]
+          }
+        ]
+      }
     }
   },
   computed: {
@@ -116,9 +92,6 @@ export default {
       }
     },
   },
-  created () {
-    this.getList()
-  },
   methods: {
     confAuthority (column) {
       return this.userType !== 3 && this.userType <= column.source
@@ -126,59 +99,42 @@ export default {
     operAuthority (column) {
       return this.userType <= column.source
     },
-    getList (page = this.page, form = this.searchForm) {
-      this.boxLoading = true
-      let formData = {
-        current: page.currentPage,
-        size: page.pageSize,
-        ...form,
-        cityId: this.userInfo.manageCityId
-      }
-      if (this.isAdmin) {
-        formData.source = 1
-      }
-      getColumnList(formData).then(({data}) => {
-        if (data.code === 0) {
-          this.columnList = data.data.data.records
-          this.page = {
-            ...page,
-            total: data.data.data.total
-          }
-        }
-      }).finally(() => {
-        this.boxLoading = false
+    fetchListFun (params) {
+      return new Promise((resolve, reject) => {
+        this.boxLoading = true
+        getColumnList({
+          ...params,
+          cityId: this.userInfo.manageCityId
+        }).then(({data}) => {
+          this.boxLoading = false
+          resolve({
+            records: data.data.data.records,
+            page: {
+              total: data.data.data.total
+            }
+          })
+        })
       })
     },
-    handleCreate () {
-      this.formData = {
+    addFun(formData, next) {
+      addColumn({
         cityIdList: [this.userInfo.manageCityId],
-        closeAllowed: 1,
-      }
-      this.formType = 'add'
-      this.formDialogVisible = true
-    },
-    create() {
-      addColumn(this.formData).then(({data}) => {
-        this.formDialogVisible = false
+        ...formData
+      }).then(({data}) => {
         this.$notify({
           title: '成功',
           message: '创建成功',
           type: 'success',
           duration: 2000
         })
-        this.page.currentPage = 1
-        this.getList()
-      }).catch(() => {
-        loading()
+        next()
       })
     },
     handleUpdate (row) {
-      this.formData = row
-      this.formType = 'edit'
-      this.formDialogVisible = true
+      this.$refs.hcCrud.rowEdit(row)
     },
-    update() {
-      updateColumn(this.formData).then(({data}) => {
+    updateFun(formData, next) {
+      updateColumn(formData).then(({data}) => {
         this.formDialogVisible = false
         this.$notify({
           title: '成功',
@@ -186,9 +142,7 @@ export default {
           type: 'success',
           duration: 2000
         })
-        this.getList()
-      }).catch(() => {
-        loading()
+        next()
       })
     },
     cityView (columnId) {
@@ -197,12 +151,18 @@ export default {
       })
     },
     handleDel (officialColumnId) {
-      deleteColumn({cityId: this.userInfo.manageCityId, officialColumnId}).then(({data}) => {
-        if (data.code === 0) {
-          this.$message.success('删除成功')
-          this.getList()
-        }
-      })
+      this.$confirm("是否删除该栏目?", "提示", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning",
+      }).then(() => {
+        deleteColumn({cityId: this.userInfo.manageCityId, officialColumnId}).then(({data}) => {
+          if (data.code === 0) {
+            this.$message.success('删除成功')
+            this.$refs.hcCrud.refresh()
+          }
+        })
+      });
     },
     handleStart (row) {
       let havEnable = row.havEnable ? 0 : 1
@@ -213,24 +173,10 @@ export default {
       }).then(({data}) => {
         if (data.code === 0) {
           this.$message.success('操作成功')
-          this.getList()
+          this.$refs.hcCrud.refresh()
         }
       })
     }, 
-    toSearch () {
-      this.searchForm = this.tempSearch
-      this.page.currentPage = 1
-      this.getList()
-    },
-    currentChange (current) {
-      this.page.currentPage = current
-      this.getList()
-    },
-    sizeChange (size) {
-      this.page.pageSize = size
-      this.page.currentPage = 1
-      this.getList()
-    }
   }
 }
 </script>
